Part two of folder restructure
Updated class files with imports for class files that are now in
different locations.
