+++ kramer.jeff.stock.option/src/com/jkramer/dao/StockDAOImpl.java
@@ -1,4 +1,4 @@
-package kramer.jeff.stock.option;
+package com.jkramer.dao;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -6,6 +6,10 @@ import java.sql.Statement;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.util.LinkedHashMap;
+
+import com.jkramer.model.Stock;
+
+import kramer.jeff.stock.option.Constants;
 
 public class StockDAOImpl implements StockDAO{
