+++ kramer.jeff.stock.option/src/com/jkramer/model/Stock.java
@@ -1,4 +1,4 @@
-package kramer.jeff.stock.option;
+package com.jkramer.model;
 
 import java.util.LinkedHashMap;
