+++ kramer.jeff.stock.option/src/kramer/jeff/stock/option/GUINewAccount.java
@@ -9,6 +9,10 @@ import javafx.scene.paint.Color;
 import javafx.scene.text.*;
 import javafx.stage.Stage;
 import java.util.TreeMap;
+
+import com.jkramer.model.Account;
+import com.jkramer.service.AccountService;
+
 import java.util.Date;
 import java.time.LocalDate;
