+++ kramer.jeff.stock.option/src/kramer/jeff/stock/option/GUINewStock.java
@@ -11,6 +11,10 @@ import javafx.stage.Stage;
 import java.util.HashMap;
 import java.util.TreeMap;
 
+import com.jkramer.model.Account;
+import com.jkramer.model.Stock;
+import com.jkramer.service.StockService;
+
 /**
  * Creates the window for the user to enter new stocks
  * @author J Kramer
