+++ kramer.jeff.stock.option/test/kramer/jeff/stock/option/PriceTest.java
@@ -2,6 +2,9 @@ package kramer.jeff.stock.option;
 
 import static org.junit.Assert.*;
 import org.junit.Test;
+
+import com.jkramer.model.Price;
+
 import junit.framework.*;
 
 import java.util.Calendar;
